Add dry-run capability to case.build

+++ scripts/lib/CIME/build.py
@@ -154,7 +154,7 @@ def _build_model(build_threaded, exeroot, incroot, complist,
 
 ###############################################################################
 def _build_model_cmake(exeroot, complist, lid, cimeroot, buildlist,
-                       comp_interface, sharedpath, use_gmake, case):
+                       comp_interface, sharedpath, use_gmake, dry_run, case):
 ###############################################################################
     cime_model = get_model()
     bldroot    = os.path.join(exeroot, "cmake-bld")
@@ -187,18 +187,31 @@ def _build_model_cmake(exeroot, complist, lid, cimeroot, buildlist,
         cmake_args += " -GNinja "
         cmake_env += "PATH={}:$PATH ".format(ninja_path)
 
-    cmake_cmd = "{}cmake {} {}/components >> {} 2>&1".format(cmake_env, cmake_args, srcroot, bldlog)
-    with open(bldlog, "w") as fd:
-        fd.write("Configuring with cmake cmd:\n{}\n\n".format(cmake_cmd))
-    stat = run_cmd(cmake_cmd, from_dir=bldroot)[0]
+    cmake_cmd = "{}cmake {} {}/components".format(cmake_env, cmake_args, srcroot)
+    stat = 0
+    if dry_run:
+        logger.info("CMake cmd:\ncd {} && {}\n\n".format(bldroot, cmake_cmd))
+    else:
+        with open(bldlog, "w") as fd:
+            fd.write("Configuring with cmake cmd:\n{}\n\n".format(cmake_cmd))
+
+        # Add logging before running
+        cmake_cmd = "{} >> {} 2>&1".format(cmake_cmd, bldlog)
+        stat = run_cmd(cmake_cmd, from_dir=bldroot)[0]
 
     # Call Make
     if stat == 0:
-        make_cmd = "{} -j {} >> {} 2>&1".format(gmake if use_gmake else "{} -v".format(os.path.join(ninja_path, "ninja")), gmake_j, bldlog)
-        with open(bldlog, "a") as fd:
-            fd.write("\n\nBuilding with cmd:\n{}\n\n".format(make_cmd))
+        make_cmd = "{} -j {}".format(gmake if use_gmake else "{} -v".format(os.path.join(ninja_path, "ninja")), gmake_j)
+        if dry_run:
+            logger.info("Build cmd:\ncd {} && {}\n\n".format(bldroot, make_cmd))
+            expect(False, "User requested dry-run only, terminating build")
+        else:
+            with open(bldlog, "a") as fd:
+                fd.write("\n\nBuilding with cmd:\n{}\n\n".format(make_cmd))
 
-        stat = run_cmd(make_cmd, from_dir=bldroot)[0]
+            # Add logging before running
+            make_cmd = "{} >> {} 2>&1".format(make_cmd, bldlog)
+            stat = run_cmd(make_cmd, from_dir=bldroot)[0]
 
     expect(stat == 0, "BUILD FAIL: build {} failed, cat {}".format(cime_model, bldlog))
 
@@ -469,13 +482,15 @@ def _clean_impl(case, cleanlist, clean_all, clean_depends):
 
 ###############################################################################
 def _case_build_impl(caseroot, case, sharedlib_only, model_only, buildlist,
-                     save_build_provenance, use_old, use_gmake):
+                     save_build_provenance, use_old, use_gmake, dry_run):
 ###############################################################################
 
     t1 = time.time()
 
     expect(not (sharedlib_only and model_only),
            "Contradiction: both sharedlib_only and model_only")
+    expect(not (dry_run and not model_only),
+           "Dry-run is only for model builds, please build sharedlibs first")
     logger.info("Building case in directory {}".format(caseroot))
     logger.info("sharedlib_only is {}".format(sharedlib_only))
     logger.info("model_only is {}".format(model_only))
@@ -592,7 +607,7 @@ def _case_build_impl(caseroot, case, sharedlib_only, model_only, buildlist,
     if not sharedlib_only:
         if get_model() == "e3sm" and not use_old:
             logs.extend(_build_model_cmake(exeroot, complist, lid, cimeroot, buildlist,
-                                           comp_interface, sharedpath, use_gmake, case))
+                                           comp_interface, sharedpath, use_gmake, dry_run, case))
         else:
             os.environ["INSTALL_SHAREDPATH"] = os.path.join(exeroot, sharedpath) # for MPAS makefile generators
             logs.extend(_build_model(build_threaded, exeroot, incroot, complist,
@@ -637,10 +652,10 @@ def post_build(case, logs, build_complete=False, save_build_provenance=True):
         lock_file("env_build.xml", caseroot=case.get_value("CASEROOT"))
 
 ###############################################################################
-def case_build(caseroot, case, sharedlib_only=False, model_only=False, buildlist=None, save_build_provenance=True, use_old=False, use_gmake=False):
+def case_build(caseroot, case, sharedlib_only=False, model_only=False, buildlist=None, save_build_provenance=True, use_old=False, use_gmake=False, dry_run=False):
 ###############################################################################
     functor = lambda: _case_build_impl(caseroot, case, sharedlib_only, model_only, buildlist,
-                                       save_build_provenance, use_old, use_gmake)
+                                       save_build_provenance, use_old, use_gmake, dry_run)
     return run_and_log_case_status(functor, "case.build", caseroot=caseroot)
 
 ###############################################################################
